refactor: Extrai classes de InternoController
